% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_region_selector.R
\name{regionSelectorApp}
\alias{regionSelectorApp}
\title{Region Selector App}
\usage{
regionSelectorApp()
}
\description{
Region Selector App
}
\examples{
\dontrun{
tdb <- tiledb::tiledb_array("tiledb://TileDB-Inc/vcf-1kg-phase3-data")
schemaViewer(tdb)
}
}
