% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwasExplorer.R
\name{gwasExplorer}
\alias{gwasExplorer}
\title{gwasExplorer}
\description{
score non-coding variants in GWAS loci to elucidate function
}
\examples{
  gwex <- gwasExplorer$new(targetGene="NDUFS2", locusName="ADAMTS4", tagSnp="rs4575098")
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{gwasExplorer$new()}}
\item \href{#method-getLinkageTable}{\code{gwasExplorer$getLinkageTable()}}
\item \href{#method-loadLinkageTable}{\code{gwasExplorer$loadLinkageTable()}}
\item \href{#method-getChromLocs}{\code{gwasExplorer$getChromLocs()}}
\item \href{#method-getEQTLSummary}{\code{gwasExplorer$getEQTLSummary()}}
\item \href{#method-getEqtlsForGene}{\code{gwasExplorer$getEqtlsForGene()}}
\item \href{#method-trenaMultiScore}{\code{gwasExplorer$trenaMultiScore()}}
\item \href{#method-clone}{\code{gwasExplorer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this [R6][R6::R6Class] class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gwasExplorer$new(targetGene, locusName, tagSnp)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{targetGene}}{character}

\item{\code{locusName}}{haracter, from GWAS studies, often meta-analyses}

\item{\code{tagSnp}}{character}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new instance of gwasExplorer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getLinkageTable"></a>}}
\if{latex}{\out{\hypertarget{method-getLinkageTable}{}}}
\subsection{Method \code{getLinkageTable()}}{
from halporeg, a table with both hg19 and hg38 coordinates
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gwasExplorer$getLinkageTable()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-loadLinkageTable"></a>}}
\if{latex}{\out{\hypertarget{method-loadLinkageTable}{}}}
\subsection{Method \code{loadLinkageTable()}}{
from halporeg, a table with both hg19 and hg38 coordinates
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gwasExplorer$loadLinkageTable(tagSnp)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tagSnp}}{character}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getChromLocs"></a>}}
\if{latex}{\out{\hypertarget{method-getChromLocs}{}}}
\subsection{Method \code{getChromLocs()}}{
hg19 and hg38 chromosomal locations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gwasExplorer$getChromLocs(geneSymbol)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getEQTLSummary"></a>}}
\if{latex}{\out{\hypertarget{method-getEQTLSummary}{}}}
\subsection{Method \code{getEQTLSummary()}}{
retrieve a summary of eQTL studies currently available from embl
return data.frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gwasExplorer$getEQTLSummary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getEqtlsForGene"></a>}}
\if{latex}{\out{\hypertarget{method-getEqtlsForGene}{}}}
\subsection{Method \code{getEqtlsForGene()}}{
from all available sources, a table with both hg19 and hg38 coordinates
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gwasExplorer$getEqtlsForGene(shoulder, eqtl.catalog.studyIDs, pval.threshold)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{shoulder}}{numeric number of bytes up- and downstream from the largst transcript}

\item{\code{eqtl.catalog.studyIDs}}{character vector,}

\item{\code{pval.threshold}}{numeric retain only eQTLs as or more significant than this}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-trenaMultiScore"></a>}}
\if{latex}{\out{\hypertarget{method-trenaMultiScore}{}}}
\subsection{Method \code{trenaMultiScore()}}{
build tms model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gwasExplorer$trenaMultiScore(tissueName, shoulder, tbl.fimo, tbl.oc)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tissueName}}{character as used by GTEx}

\item{\code{shoulder}}{numeric number of bytes up- and downstream from the largst transcript}

\item{\code{tbl.fimo}}{data.frame, typically a very large & low threshold set of fimo scores}

\item{\code{tbl.oc}}{data.frame, open chromatin from, e.g., ATAC-seq, as tissue- and cell-type specific as possible}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gwasExplorer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
